Add an option to reverse the recommendation on the threshold strategy
Previously if the metric was above 'max' we assumed the recommendation should be to scale up.
In some cases (how much of a resource is 'free' for example) it might be the 'min' number that means we should scale up.

This adds a per-metric option ('up_means_down') to the threshold strategy that will return the inverse recommendation for that metric.